+++ backend/app/Http/Controllers/MessagesController.php
@@ -6,9 +6,18 @@ use App\Models\Messages;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreMessagesRequest;
 use App\Http\Requests\UpdateMessagesRequest;
+use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Support\Facades\Gate;
 
-class MessagesController extends Controller
+class MessagesController extends Controller implements HasMiddleware
 {
+    public static function middleware()
+    {
+        return [
+            new Middleware('auth:sanctum', except: ['index', 'show'])
+            ];
+    }
     /**
      * Display a listing of the resource.
      */
@@ -23,7 +32,7 @@ class MessagesController extends Controller
     public function store(StoreMessagesRequest $request)
     {
         $fields = $request->validate();
-        $message = Messages::create($fields);
+        $message = $request->conversation()->messages()->create($fields);
         return  ['message' => $message];
     }
 
@@ -40,6 +49,7 @@ class MessagesController extends Controller
      */
     public function update(UpdateMessagesRequest $request, Messages $messages)
     {
+        Gate::authorize('modify', $messages);
         $fields = $request->validate();
 
         $messages->update($fields);
@@ -52,6 +62,7 @@ class MessagesController extends Controller
      */
     public function destroy(Messages $messages)
     {
+        Gate::authorize('modify', $messages);
         $messages->delete();
 
         return [
